src/requirements.txt
python-dotenv
psutil
pymongo
pymongo[srv]
requests
PyNaCl
bs4
html5lib
discord.py==2.4.0
wavelink==3.4.0
humanfriendly==10.0
gtts==2.3.2
numpy==1.24.1
cashfree-pg>=4.0.0
google-generativeai
pytz